Added upstreamSequence and downStreamSequence in Range.
Chenged the method prepareSequence in Range and the IntactCloner.

We try to take in total 40 amino acids upstream/downstream so 20 upstream, 20 downstream. If we can't take 20 upstream, will take more amino acids downstream and vice-versa. Of course, we can never store more amino acids than the sequence contains.

Updated the tests in RangeTest and SchemaUtilsTest